% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgphy-package.R
\docType{package}
\name{bgphy-package}
\alias{bgphy}
\alias{bgphy-package}
\title{bgphy: Bayesian Analysis of Mixed Gaussian Phylogenetic Models}
\description{
Bayesian inference of parameters of mixed Gaussian phylogenetic models with shifts under the GLInv sub-family.
}
\author{
\strong{Maintainer}: Bayu Brahmantio \email{bayubrahmantio@gmail.com}

}
\keyword{internal}
