% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgphy.R
\name{bgphy}
\alias{bgphy}
\title{Bayesian inference of mixed Gaussian phylogenetic models.}
\usage{
bgphy(model, X, nsample = 10000, parallel = TRUE)
}
\arguments{
\item{model}{An object of class \code{bgphy_model}, which is created by \code{\link[=setModel]{setModel()}} function.}

\item{X}{Measurements on tips. \code{X} needs to be a matrix of dimension \code{(1 x N)}, where \code{N} is the number of tips on the tree in \code{model$tree}.
The column names of \code{X} needs to match the names on the tips.}

\item{nsample}{Number of random draws to be made.}

\item{parallel}{Utilization of parallel processing. Defaults to \code{TRUE}.}
}
\value{
A list of class \code{mgpm_posterior} which contains:
\itemize{
\item \code{mean}:   Expected value of the posterior distribution.
\item \code{std_error}:   Estimated standard error of the posterior mean.
\item \code{std}:   Estimated standard deviation of the posterior distribution.
\item \code{quantiles}:  Estimated quantiles of marginal posterior distribution.
\item \code{Q}:   Drawn values from the proposal distribution.
\item \code{W}:   Normalized weights for each row of \code{Q}.
\item \code{loss}:   Estimated posterior predictive loss.
\item \code{ESS}:   Estimated effective sample size.
}
}
\description{
The main function to run Bayesian inference of mixed Gaussian phylogenetic models.
}
\examples{
\dontrun{
OU1 <- setModel(tree = lizardTree, regime_names = "Regime1", modeltypes = "OU")
post_OU <- bgphy(OU1,
                 matrix(XOU[1,], nrow = 1,
                        dimnames = list(NULL, colnames(XOU))))

# mixed, OU to OU
OUOU <- setModel(tree = lizardTree,
                 regime_names = c("Ancestral", "New"),
                 modeltypes = c("OU", "OU"),
                 startNodes = list(Ancestral = c(101), New = c(135)))
post_OUOU <- bgphy(BMOU,
                   matrix(XMixedOU[1,], nrow = 1,
                          dimnames = list(NULL, colnames(XMixedOU))))
}

}
