% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmgpm-package.R
\docType{package}
\name{bmgpm-package}
\alias{bmgpm}
\alias{bmgpm-package}
\title{bmgpm: Bayesian Analysis of Mixed Gaussian Phylogenetic Models}
\description{
Bayesian inference of parameters of mixed Gaussian phylogenetic models with shifts under the GLInv sub-family.
}
\author{
\strong{Maintainer}: Bayu Brahmantio \email{bayubrahmantio@gmail.com}

}
\keyword{internal}
