% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{XOU}
\alias{XMixedOU}
\title{Simulated datasets}
\format{
\code{XOU} is a matrix of size 100 x 100. Each row of was simulated on \code{lizardTree} under the \code{modelOU}.

\code{XMixedOU} is a matrix of size 100 x 100. Each row of was simulated on \code{lizardTree} under the \code{modelMixedOU}.
}
\usage{
XOU

XMixedOU
}
\description{
Two datasets simulated on the \emph{Anolis} phylogeny under different evolutionary models.
}
\keyword{datasets}
