% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lizardTree}
\alias{lizardTree}
\alias{lizardTreeMixed}
\title{Greater Antillean \emph{Anolis} tree}
\format{
An object of class \code{phylo}.

\code{lizardTreeMixed} is of class \code{PCMTree} in addition to \code{phylo}.
}
\source{
Mahler, D. Luke, et al. "Exceptional convergence on the macroevolutionary
landscape in island lizard radiations." \emph{Science} 341.6143 (2013): 292-295. \url{https://doi.org/10.1126/science.1232392}
}
\usage{
lizardTree

lizardTreeMixed
}
\description{
A phylogeny of class \code{phylo} consisting of 100 different species of genus
\emph{Anolis} inhabiting the Greater Antillean Islands. \code{lizardTreeMixed} is partitioned
into two evolutionary regimes for simulation purposes.
}
\keyword{datasets}
