% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgpm.R
\name{mgpm}
\alias{mgpm}
\title{Bayesian inference of mixed Gaussian phylogenetic models.}
\usage{
mgpm(
  model,
  X,
  tree,
  priors,
  nsample = 10000,
  initial = NULL,
  scale = 1,
  parallel = TRUE
)
}
\arguments{
\item{model}{An object of class \code{PCM}.}

\item{X}{Measurements on tips.}

\item{tree}{A phylogenetic tree of class \code{phylo}.}

\item{priors}{A prior object of class \code{mgpm_prior}.}

\item{nsample}{Number of random draws to be made.}

\item{scale}{The scale parameter to determine the scale of the approximated covariance matrix
obtained from the Laplace's approximation of the posterior distribution. The default is 1.}

\item{parallel}{Utilization of parallel processing. Defaults to \code{TRUE}.}
}
\value{
A list of class \code{mgpm_posterior} which contains:
\itemize{
\item \code{E}:   Expected value of the posterior distribution.
\item \code{Q}:   Values used to calculate the expected value using Importance Sampling.
\item \code{W}:   Normalized weights for each row of \code{Q}.
\item \code{WAIC}:   Widely Available Criterion Score.
}
}
\description{
The main function to run Bayesian inference of mixed Gaussian phylogenetic models.
}
\examples{
# OU <- PCMBase::PCM("OU")
# priorOU <- setPriors(OU)
# priorOU$X0 <- prior_normal(mean = 0, sd = 1)
# priorOU$H <- prior_halfnormal(sigma = 1)
# priorOU$Theta <- prior_normal(mean = 0, sd = 1)
# priorOU$Sigma_x <- prior_halfnormal(sigma = 1)
#
# post <- mgpm(OU, XOU[1,], lizardTree, priorOU, 100)

}
