% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{models}
\alias{models}
\alias{modelOU}
\alias{modelMixedOU}
\title{Single and Mixed Ornstein-Uhlenbeck model}
\format{
An object of class \code{PCM}.

An object of class \code{PCM}.
}
\source{
Mitov, Venelin, et al. "Fast likelihood calculation for multivariate Gaussian phylogenetic models with shifts."
\emph{Theoretical Population Biology} 131 (2020): 66-78. \url{https://doi.org/10.1016/j.tpb.2019.11.005}
}
\usage{
modelOU

modelMixedOU
}
\description{
Ornstein-Uhlenbeck (OU) models of evolution. \code{modelOU} is a single (global) model
while \code{modelMixedOU} is consisted of two OU models.
}
\keyword{datasets}
