% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{plot_regimes}
\alias{plot_regimes}
\title{Plot a tree with regimes configuration.}
\usage{
plot_regimes(tree, cols = NULL, ...)
}
\arguments{
\item{tree}{A tree of subclass \code{PCMTree}, obtained from \link{setModel} definition.}

\item{cols}{Selected color palette. If \code{NULL}, the default colors are set internally.}

\item{...}{Other arguments to be passed to \code{\link[ape:plot.phylo]{ape::plot.phylo()}}.}
}
\description{
Plots the tree object of class \code{PCMTree}, along with regimes configuration.
}
\examples{
\dontrun{
# global model
OU1 <- setModel(tree = lizardTree, regime_names = "Regime1", modeltypes = "OU")
plot(OU1$tree)

# mixed, BM to OU
BMOU <- setModel(tree = lizardTree,
                 regime_names = c("Ancestral", "New"),
                 modeltypes = c("BM", "OU"),
                 startNodes = list(Ancestral = c(101), New = c(135)))
plot_regimes(BMOU$tree)
}

}
