% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgphy.R
\name{post_pred_check}
\alias{post_pred_check}
\title{Posterior predictive check}
\usage{
post_pred_check(post, nsim = 100)
}
\arguments{
\item{post}{An object of class \code{bgphy_posterior}.}

\item{nsim}{Number of simulations, or draws from the posterior predictive distribution.}
}
\description{
Plots a number of density curves drawn from the posterior predictive distribution and the data density curve.
}
\examples{
\dontrun{
OU <- setModel(tree = lizardTree, modeltypes = ("OU"))
X <- matrix(XOU[1,], nrow = 1, dimnames = list(NULL, colnames(XOU)))
post <- bgphy(OU, X, nsample = 1000)
post_pred_check(post)
}

}
