% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgphy.R
\name{post_simulate}
\alias{post_simulate}
\title{Simulate posterior samples}
\usage{
post_simulate(post, nsample = 10000)
}
\arguments{
\item{post}{An object of class \code{bgphy_posterior}.}

\item{nsample}{Number of samples or draws from the posterior distribution.}
}
\description{
Generates sample from the posterior distribution using sampling-importance resampling.
}
\examples{
\dontrun{
OU1 <- setModel(tree = lizardTree, regime_names = "Regime1", modeltypes = "OU")
post_OU <- bgphy(OU1,
                 matrix(XOU[1,], nrow = 1,
                        dimnames = list(NULL, colnames(XOU))))
post_simulate(post_OU)

# mixed, OU to OU
OUOU <- setModel(tree = lizardTree,
                 regime_names = c("Ancestral", "New"),
                 modeltypes = c("OU", "OU"),
                 startNodes = list(Ancestral = c(101), New = c(135)))
post_OUOU <- bgphy(BMOU,
                   matrix(XMixedOU[1,], nrow = 1,
                          dimnames = list(NULL, colnames(XMixedOU))))
post_simulate(post_OUOU)
}

}
