% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{print.bgphy_model}
\alias{print.bgphy_model}
\title{Print model information}
\usage{
\method{print}{bgphy_model}(x, ...)
}
\arguments{
\item{post}{An object of class \code{bgphy_model}.}
}
\description{
This function prints the information of a model object of class \code{bgphy_model} on the console.
}
\examples{
\dontrun{
# global model
OU1 <- setModel(tree = lizardTree, regime_names = "Regime1", modeltypes = "OU")
print(OU1)

BMOU <- setModel(tree = lizardTree,
                 regime_names = c("Ancestral", "New"),
                 modeltypes = c("BM", "OU"),
                 startNodes = list(Ancestral = c(101), New = c(135)))
print(BMOU)
}

}
