% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgphy.R
\name{print.bgphy_posterior}
\alias{print.bgphy_posterior}
\title{Print posterior information}
\usage{
\method{print}{bgphy_posterior}(post, ...)
}
\arguments{
\item{post}{An object of class \code{bgphy_posterior}.}
}
\description{
This function prints the information of a posterior distribution on the console.
}
\examples{
\dontrun{
OU <- setModel(tree = lizardTree, modeltypes = ("OU"))
X <- matrix(XOU[1,], nrow = 1, dimnames = list(NULL, colnames(XOU)))
post <- bgphy(OU, X, nsample = 1000)
print(post)
}

}
