% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{print.bgphy_priors}
\alias{print.bgphy_priors}
\title{Print priors information}
\usage{
\method{print}{bgphy_priors}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bgphy_priors}.}

\item{...}{Other arguments to be passed to \code{print} or \code{print.priorpdf}}
}
\description{
Prints objects of class \code{bgphy_priors}, which contains objects of class \code{priorpdf}.
}
\examples{
\dontrun{
# global model
OU1 <- setModel(tree = lizardTree, regime_names = "Regime1", modeltypes = "OU")
OU1$priors

# mixed, BM to OU
BMOU <- setModel(tree = lizardTree,
                 regime_names = c("Ancestral", "New"),
                 modeltypes = c("BM", "OU"),
                 startNodes = list(Ancestral = c(101), New = c(135)))
BMOU$priors
}

}
