% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{print.priorpdf}
\alias{print.priorpdf}
\title{#' @export
priorPDF <- function(f, bounds){
fname <- deparse(substitute(f))
class(f) <- c("priorpdf", "custom")
attr(f, "bounds") <- bounds
attr(f, "fname") <- fname
return(f)
}}
\usage{
\method{print}{priorpdf}(prior, ...)
}
\description{
#' @export
priorPDF <- function(f, bounds){
fname <- deparse(substitute(f))
class(f) <- c("priorpdf", "custom")
attr(f, "bounds") <- bounds
attr(f, "fname") <- fname
return(f)
}
}
