% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{priorpdf}
\alias{priorpdf}
\alias{prior_uniform}
\alias{prior_normal}
\alias{prior_gamma}
\alias{prior_halfnormal}
\alias{prior_halfcauchy}
\alias{prior_halft}
\title{log Probability density functions.}
\usage{
prior_uniform(min = 0, max = 1)

prior_normal(mean = 0, sd = 1)

prior_gamma(shape, rate = 1, scale = 1/rate)

prior_halfnormal(sigma)

prior_halfcauchy(sigma)

prior_halft(nu, sigma)
}
\value{
A log probability density function of class \code{priorpdf}.
}
\description{
This is a collection of log probability density functions used for the prior object of class \code{mgpm_prior}.
}
\examples{
\dontrun{
modeltypes <- setNames(c("OU", "OU"), c("ancestral", "new"))
startNodes <- setNames(c(101, 135), c("ancestral", "new"))
OU <- setModel(tree = lizardTree, modeltypes = ("OU"))
# set custom priors
OU$priors$X0 <- prior_normal(mean = 0, sd = 2)
OU$priors$alpha <- prior_halfnormal(sigma = 3)
OU$priors$theta_1 <- prior_normal(mean = 2, sd = 2)
}

}
