% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{priorsampler}
\alias{priorsampler}
\alias{priorsampler.uniform}
\alias{priorsampler.normal}
\alias{priorsampler.gamma}
\alias{priorsampler.halfnormal}
\alias{priorsampler.halfcauchy}
\alias{priorsampler.halft}
\title{RNG samplers for prior distribution.}
\usage{
priorsampler(priorpdf)

\method{priorsampler}{uniform}(priorpdf)

\method{priorsampler}{normal}(priorpdf)

\method{priorsampler}{gamma}(priorpdf)

\method{priorsampler}{halfnormal}(priorpdf)

\method{priorsampler}{halfcauchy}(priorpdf)

\method{priorsampler}{halft}(priorpdf)
}
\arguments{
\item{priorpdf}{An object of class \code{priorpdf}.}
}
\value{
A sampler of class \code{sampler} with subclass that depends on the type
of distribution of the \code{priorpdf} object. This sampler only takes number
of samples \code{n} as its argument, while the distributional parameters are
fully specified by the \code{priorpdf} object.
}
\description{
Converts a \code{priorpdf} object into an RNG sampler.
}
\examples{
\dontrun{
OU1 <- setModel(tree = lizardTree, regime_names = "Regime1", modeltypes = "OU")
ralpha <- priorsampler(OU1$priors$alpha)
ralpha(100)
}

}
