% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{setModel}
\alias{setModel}
\title{Set a Gaussian phylogenetic model}
\usage{
setModel(tree, modeltypes, startNodes = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{phylo}.}

\item{modeltypes}{A named vector of model types. The elements define the model types and names define the regime names.
Currently only \code{"BM"} and \code{"OU"} are available as model types.}

\item{startNodes}{A named vector of nodes. Each element determines the start of a regime on the tree.}
}
\value{
An object of class \code{bgphy_model}, which contains:
\itemize{
\item \code{model}: The PCM model as defined by \link{PCMBase}.
\item \code{tree}: The phylogenetic tree. If there are multiple regimes, the tree is converted into \code{PCMTree} class.
\item \code{priors}: A list of default priors. Each of them is of class \code{priorpdf}.
}
}
\description{
Defines a Gaussian phylogenetic model.
}
\examples{
\dontrun{
modeltypes <- setNames(c("OU", "OU"), c("ancestral", "new"))
startNodes <- setNames(c(101, 135), c("ancestral", "new"))
OU <- setModel(tree = lizardTree, modeltypes = ("OU"))
}

}
