% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{setPriors}
\alias{setPriors}
\title{Set a prior object given a model of class \code{PCM}.}
\usage{
setPriors(model)
}
\arguments{
\item{model}{A \code{PCM} object.}
}
\value{
A prior object of class \code{mgpm_prior}, which is required for the main function, \code{mgpm}.
}
\description{
This function takes a model of class \code{PCM}, created from \code{PCMBase},
and outputs a blank prior object of class \code{mgpm_prior}.
Users can then modify the prior accordingly using functions of class \code{priorpdf}.
Modifying an \code{mgpm_prior} object is similar to modifying a list,
in which the users can replace the values in the list with \code{priorpdf} functions
provided by the \code{bmgpm} package or by the users.
}
\examples{
OU <- PCMBase::PCM("OU")
priorOU <- setPriors(OU)
priorOU$X0 <- prior_normal(mean = 0, sd = 1)
priorOU$H <- prior_halfnormal(sigma = 1)
priorOU$Theta <- prior_normal(mean = 0, sd = 1)
priorOU$Sigma_x <- prior_halfnormal(sigma = 1)
}
