% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgphy.R
\name{post_pred_check}
\alias{post_pred_check}
\title{Posterior predictive check}
\usage{
post_pred_check(post, ...)
}
\arguments{
\item{post}{An object of class \code{bgphy_posterior}.}

\item{...}{Other arguments to be passed on to the \code{plot} function.}
}
\description{
Plots a number of density curves drawn from the posterior predictive distribution and the data density curve.
}
\examples{
\dontrun{
OU1 <- setModel(tree = lizardTree, regime_names = "Regime1", modeltypes = "OU")
post_OU <- bgphy(OU1,
                 matrix(XOU[1,], nrow = 1,
                        dimnames = list(NULL, colnames(XOU))))
post_pred_check(post_OU)

# mixed, OU to OU
OUOU <- setModel(tree = lizardTree,
                 regime_names = c("Ancestral", "New"),
                 modeltypes = c("OU", "OU"),
                 startNodes = list(Ancestral = c(101), New = c(135)))
post_OUOU <- bgphy(BMOU,
                   matrix(XMixedOU[1,], nrow = 1,
                          dimnames = list(NULL, colnames(XMixedOU))))
post_pred_check(post_OUOU)
}

}
