% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{print.bgphy_model}
\alias{print.bgphy_model}
\title{Print model information}
\usage{
\method{print}{bgphy_model}(model)
}
\arguments{
\item{post}{An object of class \code{bgphy_model}.}
}
\description{
This function prints the information of a model object of class \code{bgphy_model} on the console.
}
\examples{
\dontrun{
modeltypes <- setNames(c("OU", "OU"), c("ancestral", "new"))
startNodes <- setNames(c(101, 135), c("ancestral", "new"))
OU <- setModel(tree = lizardTree, modeltypes = ("OU"))
print(OU)
}

}
