% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgphy.R
\name{print.bgphy_posterior}
\alias{print.bgphy_posterior}
\title{Print posterior information}
\usage{
\method{print}{bgphy_posterior}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bgphy_posterior}.}

\item{...}{Further argument to be passed to \code{print}.}
}
\description{
This function prints the information of a posterior distribution on the console.
}
\examples{
\dontrun{
# global model
OU1 <- setModel(tree = lizardTree, regime_names = "Regime1", modeltypes = "OU")
post_OU <- bgphy(OU1,
                 matrix(XOU[1,], nrow = 1,
                        dimnames = list(NULL, colnames(XOU))))
print(post_OU)

# mixed, OU to OU
OUOU <- setModel(tree = lizardTree,
                 regime_names = c("Ancestral", "New"),
                 modeltypes = c("OU", "OU"),
                 startNodes = list(Ancestral = c(101), New = c(135)))
post_OUOU <- bgphy(BMOU,
                   matrix(XMixedOU[1,], nrow = 1,
                          dimnames = list(NULL, colnames(XMixedOU))))
print(post_OUOU)
}

}
