% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{priorpdf}
\alias{priorpdf}
\alias{prior_uniform}
\alias{prior_normal}
\alias{prior_gamma}
\alias{prior_halfnormal}
\alias{prior_halfcauchy}
\alias{prior_halft}
\title{log Probability density functions.}
\usage{
prior_uniform(min = 0, max = 1)

prior_normal(mean = 0, sd = 1)

prior_gamma(shape, rate = 1, scale = 1/rate)

prior_halfnormal(sigma)

prior_halfcauchy(sigma)

prior_halft(nu, sigma)
}
\value{
A log probability density function of class \code{priorpdf}.
}
\description{
This is a collection of log probability density functions used for the prior object of class \code{bgphy_priors}.
}
\examples{
\dontrun{
# global model
OU1 <- setModel(tree = lizardTree, regime_names = "Regime1", modeltypes = "OU")
# set custom priors
OU1$priors$X0 <- OU1$priors$theta <- prior_normal(mean = 0, sd = 3)
OU1$priors$alpha <- OU1$priors$sigma <- prior_halfnormal(sigma = 2)

# mixed, BM to OU
BMOU <- setModel(tree = lizardTree,
                 regime_names = c("Ancestral", "New"),
                 modeltypes = c("BM", "OU"),
                 startNodes = list(Ancestral = c(101), New = c(135)))
# set custom priors
BMOU$priors$X0 <- prior_normal(mean = 0, sd = 3)
BMOU$priors$sigma_1 <- prior_halft(nu = 1, sigma = 2)
BMOU$priors$alpha_2 <- prior_halfnormal(sigma = 5)
BMOU$priors$theta_2 <- prior_normal(mean = 3, sd = 5)
BMOU$priors$sigma_2 <- prior_halfnormal(sigma = 3)
}

}
