% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{priorpdf}
\alias{priorpdf}
\alias{prior_uniform}
\alias{prior_normal}
\alias{prior_gamma}
\alias{prior_halfnormal}
\alias{prior_halfcauchy}
\title{log Probability density functions.}
\usage{
prior_uniform(min = 0, max = 1)

prior_normal(mean = 0, sd = 1)

prior_gamma(shape, rate = 1, scale = 1/rate)

prior_halfnormal(sigma)

prior_halfcauchy(sigma)
}
\value{
A log probability density function of class \code{priorpdf}.
}
\description{
This is a collection of log probability density functions used for the prior object of class \code{mgpm_prior}.
}
\examples{
OU <- PCMBase::PCM("OU")
priorOU <- setPriors(OU)
priorOU$X0 <- prior_normal(mean = 0, sd = 1)
priorOU$H <- prior_halfnormal(sigma = 1)
priorOU$Theta <- prior_normal(mean = 0, sd = 1)
priorOU$Sigma_x <- prior_halfnormal(sigma = 1)

}
