% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{priorsampler}
\alias{priorsampler}
\alias{priorsampler.uniform}
\alias{priorsampler.normal}
\alias{priorsampler.gamma}
\alias{priorsampler.halfnormal}
\alias{priorsampler.halfcauchy}
\alias{priorsampler.halft}
\title{RNG samplers for prior distribution.}
\usage{
priorsampler(priorpdf)

\method{priorsampler}{uniform}(priorpdf)

\method{priorsampler}{normal}(priorpdf)

\method{priorsampler}{gamma}(priorpdf)

\method{priorsampler}{halfnormal}(priorpdf)

\method{priorsampler}{halfcauchy}(priorpdf)

\method{priorsampler}{halft}(priorpdf)
}
\value{
A sampler of class \code{sampler} with subclass that depends on the type of distribution.
}
\description{
A collection of RNG sampler used for the prior object of class \code{bgphy_priors}
}
\examples{
\dontrun{
}

}
