% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{setModel}
\alias{setModel}
\title{Set a Gaussian phylogenetic model}
\usage{
setModel(tree, regime_names, modeltypes, startNodes = NULL)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{phylo}.}

\item{regime_names}{A character vector of regime names.}

\item{modeltypes}{A vector of model types. Currently only \code{"BM"} and \code{"OU"} are available as model types.}

\item{startNodes}{A list of nodes that determine the start of regimes. The names of elements in the list
must match the names in the \code{regime_names}. Does not need to be specified if there is only one (global) regime.}
}
\value{
An object of class \code{bgphy_model}, which contains:
\itemize{
\item \code{model}: The PCM model as defined by \code{\link[PCMBase:PCM]{PCMBase::PCM()}}.
\item \code{tree}: The phylogenetic tree. If there are multiple regimes, the tree is converted into \code{PCMTree} class.
\item \code{priors}: A list of default priors. Each of them is of class \code{priorpdf}.
}
}
\description{
Defines a Gaussian phylogenetic model.
}
\examples{
\dontrun{
# global model
OU1 <- setModel(tree = lizardTree, regime_names = "Regime1", modeltypes = "OU")

BMOU <- setModel(tree = lizardTree,
                 regime_names = c("Ancestral", "New"),
                 modeltypes = c("BM", "OU"),
                 startNodes = list(Ancestral = c(101), New = c(135)))
}

}
